[Ram] Changed folder structure for pages
