--- client/src/pages/Animals/index.ts
@@ -1,3 +1,0 @@
-import AnimalsTable from './AnimalsTable';
-
-export { AnimalsTable };
