--- client/src/pages/DeleteAnimals/index.ts
@@ -1,3 +1,0 @@
-import AnimalDelete from './Card';
-
-export { AnimalDelete };
