--- client/src/pages/DeletePlants/index.ts
@@ -1,3 +1,0 @@
-import PlantDelete from './Card';
-
-export { PlantDelete };
