--- client/src/pages/Plants/index.ts
@@ -1,4 +1,0 @@
-import PlantView from './PlantView';
-import PlantsTable from './PlantsTable';
-
-export { PlantsTable, PlantView };
