+++ client/src/pages/index.ts
@@ -1,11 +1,23 @@
+import AnimalDelete from './AnimalDelete';
+import AnimalsTable from './AnimalsTable';
 import Home from './Home';
 import Login from './Login';
 import PageNotFound from './PageNotFound';
+import PlantDelete from './PlantDelete';
+import PlantView from './PlantView';
+import PlantsTable from './PlantsTable';
 
-export * from './DeleteAnimals';
-export * from './DeletePlants';
-export * from './Plants';
-export * from './Animals';
+export * from './AnimalDelete';
+export * from './PlantDelete';
 export * from './CreateRecord';
 
-export { Home, Login, PageNotFound };
+export {
+  AnimalDelete,
+  AnimalsTable,
+  Home,
+  Login,
+  PageNotFound,
+  PlantDelete,
+  PlantView,
+  PlantsTable,
+};
